rag-service/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
langchain
langchain-community
langchain-openai
openai
tiktoken
pypdf
